server/finchat/api/views.py
from rest_framework.decorators import api_view
from rest_framework.response import Response
# Create your views here.

@api_view(['POST'])
def process_api(request):
    # Assuming you expect some data in the request
    data = request.data
    
    # Process the data as needed
    processed_data = process_data(data)
    
    # Return the processed data in the response
    return Response({'result': processed_data})

def process_data(data):
    # Your custom data processing logic goes here
    # For example, you might just return the input data for simplicity
    return data
